fix(capabilities): record Seed-side audit events and seed D1 rows verbatim
Twelve mutation paths recorded audit events in Live but nothing in Seed,
violating the package's own invariant; the membership, lifecycle,
invitation, and user-admin Seed adapters now record the same events, and
their contract case lists assert audit through AuditEventLog so both
adapters are held to parity mechanically.

scripts/seed.ts wrote each audit row's own id as targetId and rebuilt
actorUserId by display-name matching, silently nulling rows the fixture
carries explicitly; both columns now come from the fixture source rows.

Also: CapabilityBindings single record (7 wiring sites -> 2), the 3x
endpoint aggregate select deduped, platform-user-admin split per the
documented seam with membership and audit-log Live halves extracted,
shared withTraceparent/bestEffort/WEBHOOK_USER_AGENT/DELIVERIES_PAGE_SIZE
constants, deliveries total-order tie-break with the contract assertion
strengthened, and billing's env gate on hasValue.
