+++ packages/capabilities/package.json
@@ -26,6 +26,7 @@
   "dependencies": {
     "@b2b-saas-starter/authz": "workspace:*",
     "@b2b-saas-starter/db": "workspace:*",
+    "@b2b-saas-starter/env": "workspace:*",
     "@b2b-saas-starter/failure": "workspace:*",
     "@b2b-saas-starter/logger": "workspace:*",
     "drizzle-orm": "catalog:",
