+++ packages/capabilities/src/governance/workspace-lifecycle.contract.ts
@@ -2,7 +2,8 @@ import { Effect } from 'effect'
 import { type ContractExpectMatchers } from './contract-expect.ts'
 import { type CapabilityUnavailable, type WorkspaceChangeRejected } from '../errors.ts'
 import { failureTag } from '../internal/failure-tag.ts'
-import { type WorkspaceContext } from '../workspace-context.ts'
+import { WorkspaceContext } from '../workspace-context.ts'
+import { AuditEventLog } from './audit-event-log.ts'
 import { WorkspaceLifecycle } from './workspace-lifecycle.ts'
 
 /**
@@ -29,7 +30,7 @@ export type LifecycleContractCase = {
   readonly assert: Effect.Effect<
     void,
     CapabilityUnavailable | WorkspaceChangeRejected,
-    WorkspaceLifecycle | WorkspaceContext
+    WorkspaceLifecycle | WorkspaceContext | AuditEventLog
   >
 }
 
@@ -79,6 +80,56 @@ export function workspaceLifecycleContractCases(
         expect(renamed.name).toBe('Renamed Lab')
         // Everything else about the workspace survives the rename untouched.
         expect(renamed.slug).toContain('-lab')
+      })
+    },
+    {
+      // One audit case for the mutation family that both adapters can share
+      // (delete is deliberately absent below, as its note explains). Same
+      // delta-around-my-own-mutations shape as the developer-platform
+      // contract's `api_token.revoked` case.
+      name: 'creating and renaming a workspace records workspace audit events',
+      assert: Effect.gen(function* () {
+        const lifecycle = yield* WorkspaceLifecycle
+        const log = yield* AuditEventLog
+        const ctx = yield* WorkspaceContext
+        function globalCountOf(eventType: string) {
+          return Effect.map(
+            log.listGlobal,
+            (events) => events.filter((event) => event.eventType === eventType).length
+          )
+        }
+
+        // `create` is identity-keyed, so its event is attributed to the
+        // workspace being created — readable globally, not from this
+        // context's per-workspace trail.
+        const createdBefore = yield* globalCountOf('workspace.created')
+        const created = yield* lifecycle.create({
+          name: 'Audited Lab',
+          slug: `${ids.existingSlug}-audited`,
+          userId: ids.creator
+        })
+        const createdEvents = (yield* log.listGlobal).filter(
+          (event) => event.eventType === 'workspace.created'
+        )
+        expect(createdEvents.length).toBe(createdBefore + 1)
+        expect(
+          createdEvents.some(
+            (event) => event.targetId === created.id && event.targetType === 'workspace'
+          )
+        ).toBe(true)
+
+        // The rename acts on the workspace in context, so its event lands in
+        // this context's own trail.
+        const renamedBefore = yield* globalCountOf('workspace.renamed')
+        yield* lifecycle.rename({ name: 'Audited Lab II' })
+        const renamedPage = yield* log.list({ eventType: 'workspace.renamed' })
+        expect(renamedPage.events.length).toBe(renamedBefore + 1)
+        expect(
+          renamedPage.events.some(
+            (event) =>
+              event.targetId === ctx.workspace.id && event.targetType === 'workspace'
+          )
+        ).toBe(true)
       })
     }
     // Deletion is deliberately absent from the shared list: `remove` acts on
