+++ packages/capabilities/src/governance/workspace-sso-connections.ts
@@ -68,7 +68,7 @@ export type CreateOidcConnectionInput = {
   /**
    * Resolved by the app's test/discovery step before the connection is
    * registered, so the plugin stores a fully hydrated config and never needs
-   * runtime discovery at sign-in (ADR 0055: the strict `trustedOrigins` check
+   * runtime discovery at sign-in (ADR 0069: the strict `trustedOrigins` check
    * the plugin applies to discovery URLs would otherwise make every IdP an
    * operator env change).
    */
@@ -211,7 +211,7 @@ export type SsoConnectionsInterface = {
   >
 
   /**
-   * The domain-routing rule (ADR 0055): the connection an email address
+   * The domain-routing rule (ADR 0069): the connection an email address
    * resolves to, if any. Only `enabled` connections route, and a disabled
    * connection never intercepts a sign-in even though its row persists —
    * which is what makes the seeded example connection safe.
@@ -232,7 +232,7 @@ export type SsoConnectionsInterface = {
    * serves. `resolveRouting` is this filtered to enabled rows; the auth gate
    * in `apps/web/src/lib/server/sso-sign-in-gate.ts` reads it raw and refuses
    * a disabled answer, so "a disabled connection never intercepts sign-ins"
-   * holds at the boundary and not just on the page (ADR 0055 §2).
+   * holds at the boundary and not just on the page (ADR 0069 §2).
    */
   readonly resolveSignInTarget: (input: {
     readonly email?: string | undefined
